TGP-1267: refactor updateRecord
