+++ test/uk/gov/hmrc/tradergoodsprofilesrouter/service/RouterServiceSpec.scala
@@ -28,8 +28,6 @@ import play.api.libs.json.Json
 import play.api.mvc.Results.{BadRequest, InternalServerError}
 import uk.gov.hmrc.http.HeaderCarrier
 import uk.gov.hmrc.tradergoodsprofilesrouter.connectors.EISConnector
-import uk.gov.hmrc.tradergoodsprofilesrouter.models.request.UpdateRecordRequest
-import uk.gov.hmrc.tradergoodsprofilesrouter.models.response.CreateOrUpdateRecordResponse
 import uk.gov.hmrc.tradergoodsprofilesrouter.models.response.eis.GetEisRecordsResponse
 import uk.gov.hmrc.tradergoodsprofilesrouter.support.GetRecordsDataSupport
 
@@ -209,125 +207,5 @@ class RouterServiceSpec
       }
     }
   }
-
-  "updateRecord" should {
-    "update a record item" in {
-      val eisResponse = createOrUpdateRecordResponseData
-      when(eisConnector.updateRecord(any, any)(any))
-        .thenReturn(Future.successful(Right(eisResponse)))
-
-      val result = routerService.updateRecord(eoriNumber, "recordId", updateRecordRequest)
-
-      whenReady(result.value) {
-        _.value shouldBe eisResponse
-      }
-    }
-
-    "return an internal server error" when {
-
-      "EIS return an error" in {
-        when(eisConnector.updateRecord(any, any)(any))
-          .thenReturn(Future.successful(Left(BadRequest("error"))))
-
-        val result = routerService.updateRecord(eoriNumber, "recordId", updateRecordRequest)
-
-        whenReady(result.value) {
-          _.left.value shouldBe BadRequest("error")
-        }
-      }
-
-      "error when an exception is thrown" in {
-        when(eisConnector.updateRecord(any, any)(any))
-          .thenReturn(Future.failed(new RuntimeException("error")))
-
-        val result = routerService.updateRecord(eoriNumber, "recordId", updateRecordRequest)
-
-        whenReady(result.value) {
-          _.left.value shouldBe InternalServerError(
-            Json.obj(
-              "correlationId" -> correlationId,
-              "code"          -> "UNEXPECTED_ERROR",
-              "message"       -> "error"
-            )
-          )
-        }
-      }
-
-    }
-  }
-
-  val createOrUpdateRecordResponseData: CreateOrUpdateRecordResponse =
-    Json
-      .parse("""
-          |{
-          |  "recordId": "b2fa315b-2d31-4629-90fc-a7b1a5119873",
-          |  "eori": "GB123456789012",
-          |  "actorId": "GB098765432112",
-          |  "traderRef": "BAN001001",
-          |  "comcode": "10410100",
-          |  "accreditationStatus": "Not Requested",
-          |  "goodsDescription": "Organic bananas",
-          |  "countryOfOrigin": "EC",
-          |  "category": 1,
-          |  "supplementaryUnit": 500,
-          |  "measurementUnit": "Square metre (m2)",
-          |  "comcodeEffectiveFromDate": "2024-11-18T23:20:19Z",
-          |  "comcodeEffectiveToDate": "2024-11-18T23:20:19Z",
-          |  "version": 1,
-          |  "active": true,
-          |  "toReview": false,
-          |  "reviewReason": "Commodity code change",
-          |  "declarable": "SPIMM",
-          |  "ukimsNumber": "XIUKIM47699357400020231115081800",
-          |  "nirmsNumber": "RMS-GB-123456",
-          |  "niphlNumber": "6 S12345",
-          |  "createdDateTime": "2024-11-18T23:20:19Z",
-          |  "updatedDateTime": "2024-11-18T23:20:19Z",
-          |  "assessments": [
-          |    {
-          |      "assessmentId": "abc123",
-          |      "primaryCategory": 1,
-          |      "condition": {
-          |        "type": "abc123",
-          |        "conditionId": "Y923",
-          |        "conditionDescription": "Products not considered as waste according to Regulation (EC) No 1013/2006 as retained in UK law",
-          |        "conditionTraderText": "Excluded product"
-          |      }
-          |    }
-          |  ]
-          |}
-          |""".stripMargin)
-      .as[CreateOrUpdateRecordResponse]
-
-  val updateRecordRequest: UpdateRecordRequest = Json
-    .parse("""
-        |{
-        |    "eori": "GB123456789001",
-        |    "actorId": "GB098765432112",
-        |    "recordId": "8ebb6b04-6ab0-4fe2-ad62-e6389a8a204f",
-        |    "traderRef": "BAN001001",
-        |    "comcode": "10410100",
-        |    "goodsDescription": "Organic bananas",
-        |    "countryOfOrigin": "EC",
-        |    "category": 1,
-        |    "assessments": [
-        |        {
-        |            "assessmentId": "abc123",
-        |            "primaryCategory": 1,
-        |            "condition": {
-        |                "type": "abc123",
-        |                "conditionId": "Y923",
-        |                "conditionDescription": "Products not considered as waste according to Regulation (EC) No 1013/2006 as retained in UK law",
-        |                "conditionTraderText": "Excluded product"
-        |            }
-        |        }
-        |    ],
-        |    "supplementaryUnit": 500,
-        |    "measurementUnit": "Square metre (m2)",
-        |    "comcodeEffectiveFromDate": "2024-11-18T23:20:19Z",
-        |    "comcodeEffectiveToDate": "2024-11-18T23:20:19Z"
-        |}
-        |""".stripMargin)
-    .as[UpdateRecordRequest]
 
 }
